+++ samples/sample-jvm/src/main/kotlin/Main.kt
@@ -8,7 +8,7 @@
  * repository.
  *
  * Copyright (C) 2025-present Benoit 'BoD' Lubek (BoD@JRAF.org)
- * and contributors (https://github.com/BoD/klibfitbit/graphs/contributors)
+ * and contributors (https://github.com/BoD/klibghealth/graphs/contributors)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,62 +25,61 @@
 
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
-import org.jraf.klibfitbit.client.FitbitClient
-import org.jraf.klibfitbit.client.configuration.ClientConfiguration
-import org.jraf.klibfitbit.client.configuration.HttpConfiguration
-import org.jraf.klibfitbit.client.configuration.HttpLoggingLevel
-import org.jraf.klibfitbit.client.configuration.OAuthTokens
-import org.jraf.klibfitbit.model.ExerciseType
+import org.jraf.klibghealth.client.GoogleHealthClient
+import org.jraf.klibghealth.client.GoogleHealthClient.Configuration.Auth
+import org.jraf.klibghealth.client.GoogleHealthClient.Configuration.Auth.OAuthTokens
+import org.jraf.klibghealth.client.GoogleHealthClient.Configuration.Auth.Scope
+import org.jraf.klibghealth.model.ExerciseType
 import org.jraf.klibnanolog.logd
 import kotlin.time.Clock
 import kotlin.time.Duration.Companion.minutes
 
 suspend fun main(av: Array<String>) {
-  FitbitClient.newInstance(
-    ClientConfiguration(
-      clientId = av[0],
-      clientSecret = av[1],
-      // You can pass null the first time when you don't have OAuth tokens yet.
-      // Otherwise, pass the accessToken/refreshToken that are logged below
-      oAuthTokens = OAuthTokens(
-        accessToken = av[2],
-        refreshToken = av[3],
+  GoogleHealthClient(
+    GoogleHealthClient.Configuration(
+      Auth(
+        clientId = av[0],
+        clientSecret = av[1],
+        // You can pass null the first time when you don't have OAuth tokens yet.
+        // Otherwise, pass the accessToken/refreshToken that are logged below.
+        oAuthTokens = OAuthTokens(
+          accessToken = av[2],
+          refreshToken = av[3],
+        ),
       ),
-      httpConfiguration = HttpConfiguration(
-        loggingLevel = HttpLoggingLevel.ALL,
+      GoogleHealthClient.Configuration.Http(
+        loggingLevel = GoogleHealthClient.Configuration.Http.HttpLoggingLevel.ALL,
       ),
     ),
-  ) { oAuthTokens ->
-    logd("accessToken: " + oAuthTokens.accessToken)
-    logd("refreshToken: " + oAuthTokens.refreshToken)
-  }.use { fitbitClient ->
+  ) { newOAuthTokens ->
+    logd("accessToken: " + newOAuthTokens.accessToken)
+    logd("refreshToken: " + newOAuthTokens.refreshToken)
+  }.use { googleHealthClient ->
     // Do this only the first time, to get OAuth tokens
     val fetchTokens = false
     if (fetchTokens) {
-      val authorizationUrlResult = fitbitClient.oAuthCreateAuthorizationUrl(
-        listOf(
-          "https://www.googleapis.com/auth/googlehealth.activity_and_fitness.readonly",
-          "https://www.googleapis.com/auth/googlehealth.activity_and_fitness.writeonly",
-          "https://www.googleapis.com/auth/googlehealth.sleep.readonly",
-          "https://www.googleapis.com/auth/googlehealth.sleep.writeonly",
-        ),
+      val authorizationUrlAndCodeVerifier = googleHealthClient.oAuth.createAuthorizationUrl(
+        Scope.ActivityAndFitness.ReadOnly,
+        Scope.ActivityAndFitness.WriteOnly,
+        Scope.Sleep.ReadOnly,
+        Scope.Sleep.WriteOnly,
       )
-      println("Please visit this URL: ${authorizationUrlResult.authorizeUrl}")
+      println("Please visit this URL: ${authorizationUrlAndCodeVerifier.authorizeUrl}")
       println("Enter the callback URL:")
       val callbackUrl = readln().trim()
-      fitbitClient.oAuthFetchTokens(authorizationUrlResult, callbackUrl)
+      googleHealthClient.oAuth.fetchTokens(authorizationUrlAndCodeVerifier, callbackUrl).getOrThrow()
     }
 
     // Create new exercise data point
-    fitbitClient.createDataPoint(
-      exerciseType = ExerciseType.TREADMILL_WALK,
+    googleHealthClient.dataPoint.createDataPoint(
+      exerciseType = ExerciseType.TreadmillWalk,
       startTime = (Clock.System.now() - 5.minutes),
       activeDuration = 3.minutes,
       distanceMeters = 342.5,
-    )
+    ).getOrThrow()
 
     // Get all activities from today
     val today = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault()).date
-    logd(fitbitClient.getDataPointList(today))
+    logd(googleHealthClient.dataPoint.getDataPointList(today).getOrThrow())
   }
 }