+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Next version (unreleased)
 
+# v1.1.0 - 2026-09-26
+
+- [Breaking] This library has now been migrated to the new Google Health API. The library has been renamed from `KLibFitbit` to
+  `KLibGHealth` and the package names, Maven coordinates (`org.jraf.klibfitbit` -> `org.jraf.klibghealth`), and class names
+  (`FitbitClient` -> `GoogleHealthClient`) also renamed to reflect this change.
+
 # v1.0.2 - 2026-06-13
 
 - Enable `macosArm64` target.